Calculate profit from stocks
Say you have an array for which the ith element is the price of a given
stock on day i. If you were only permitted to complete at most one
transaction (ie, buy one and sell one share of the stock), design an
algorithm to find the maximum profit
e.g: input array [1,2,3,1,2,7,2] then o/p should be 6. Buy at 1 and sale
at 7 so 7-1 = 6